+++ packages/lext/src/components/mics/Sheet/SheetLine.tsx
@@ -20,6 +20,8 @@ const SheetLine = () => {
       style={{
         borderRadius: theme.defaultOptions.borderRadius / 2,
         width: "100%",
+        height: 40,
+        justifyContent: "center",
       }}
     >
       <Animated.View style={lineCss} />